+++ R/content_downloader.R
@@ -1,7 +1,16 @@
 content_downloader <- function(riingo_url, ticker) {
   riingo_headers <- retrieve_headers()
 
-  resp <- httr::GET(riingo_url, riingo_headers)
+  # Handled by `assert_valid_response()`
+  known_failure_status <- c(404, 400)
+
+  resp <- httr::RETRY(
+    verb = "GET",
+    url = riingo_url,
+    config = riingo_headers,
+    times = 3,
+    terminate_on = known_failure_status
+  )
 
   assert_valid_response(ticker, resp)